fix: address deferred shell review feedback
