+++ tests/helpers/runtime-config-panel-harness.mjs
@@ -4,523 +4,13 @@ import { tmpdir } from 'node:os';
 import { dirname, join, resolve } from 'node:path';
 import { fileURLToPath, pathToFileURL } from 'node:url';
 
+import { createBrowserEnvironment } from './mini-dom.mts';
+
+export { createBrowserEnvironment };
+
 const __dirname = dirname(fileURLToPath(import.meta.url));
 const root = resolve(__dirname, '..', '..');
 const entry = resolve(root, 'src/components/RuntimeConfigPanel.ts');
-
-class MiniClassList {
-  constructor() {
-    this.values = new Set();
-  }
-
-  add(...tokens) {
-    tokens.forEach((token) => this.values.add(token));
-  }
-
-  remove(...tokens) {
-    tokens.forEach((token) => this.values.delete(token));
-  }
-
-  contains(token) {
-    return this.values.has(token);
-  }
-
-  toggle(token, force) {
-    if (force === true) {
-      this.values.add(token);
-      return true;
-    }
-    if (force === false) {
-      this.values.delete(token);
-      return false;
-    }
-    if (this.values.has(token)) {
-      this.values.delete(token);
-      return false;
-    }
-    this.values.add(token);
-    return true;
-  }
-
-  setFromString(value) {
-    this.values = new Set(String(value).split(/\s+/).filter(Boolean));
-  }
-
-  toString() {
-    return Array.from(this.values).join(' ');
-  }
-}
-
-class MiniNode extends EventTarget {
-  constructor() {
-    super();
-    this.childNodes = [];
-    this.parentNode = null;
-    this.parentElement = null;
-  }
-
-  appendChild(child) {
-    if (child instanceof MiniDocumentFragment) {
-      const children = [...child.childNodes];
-      children.forEach((node) => this.appendChild(node));
-      return child;
-    }
-    if (child.parentNode) {
-      child.parentNode.removeChild(child);
-    }
-    child.parentNode = this;
-    child.parentElement = this instanceof MiniElement ? this : null;
-    this.childNodes.push(child);
-    return child;
-  }
-
-  append(...children) {
-    children.forEach((child) => {
-      if (child == null) return;
-      if (typeof child === 'string' || typeof child === 'number') {
-        this.appendChild(new MiniText(child));
-        return;
-      }
-      this.appendChild(child);
-    });
-  }
-
-  removeChild(child) {
-    const index = this.childNodes.indexOf(child);
-    if (index >= 0) {
-      this.childNodes.splice(index, 1);
-      child.parentNode = null;
-      child.parentElement = null;
-    }
-    return child;
-  }
-
-  insertBefore(child, referenceNode) {
-    if (referenceNode == null) {
-      return this.appendChild(child);
-    }
-    if (child.parentNode) {
-      child.parentNode.removeChild(child);
-    }
-    const index = this.childNodes.indexOf(referenceNode);
-    if (index === -1) {
-      return this.appendChild(child);
-    }
-    child.parentNode = this;
-    child.parentElement = this instanceof MiniElement ? this : null;
-    this.childNodes.splice(index, 0, child);
-    return child;
-  }
-
-  get firstChild() {
-    return this.childNodes[0] ?? null;
-  }
-
-  get lastChild() {
-    return this.childNodes.at(-1) ?? null;
-  }
-
-  get firstElementChild() {
-    return this.childNodes.find((child) => child instanceof MiniElement) ?? null;
-  }
-
-  get lastElementChild() {
-    return [...this.childNodes].reverse().find((child) => child instanceof MiniElement) ?? null;
-  }
-
-  get childElementCount() {
-    return this.childNodes.filter((child) => child instanceof MiniElement).length;
-  }
-
-  get textContent() {
-    return this.childNodes.map((child) => child.textContent ?? '').join('');
-  }
-
-  set textContent(value) {
-    this.childNodes = [new MiniText(value ?? '')];
-  }
-
-  replaceChildren(...children) {
-    this.childNodes = [];
-    this.append(...children);
-  }
-}
-
-class MiniText extends MiniNode {
-  constructor(value) {
-    super();
-    this.value = String(value);
-  }
-
-  get textContent() {
-    return this.value;
-  }
-
-  set textContent(value) {
-    this.value = String(value);
-  }
-
-  get outerHTML() {
-    return this.value;
-  }
-}
-
-class MiniDocumentFragment extends MiniNode {
-  get outerHTML() {
-    return this.childNodes.map((child) => child.outerHTML ?? child.textContent ?? '').join('');
-  }
-}
-
-class MiniElement extends MiniNode {
-  constructor(tagName) {
-    super();
-    this.tagName = tagName.toUpperCase();
-    this.attributes = new Map();
-    this.classList = new MiniClassList();
-    this.dataset = {};
-    this.style = {};
-    this._innerHTML = '';
-    this.id = '';
-    this.title = '';
-    this.disabled = false;
-  }
-
-  get className() {
-    return this.classList.toString();
-  }
-
-  set className(value) {
-    this.classList.setFromString(value);
-  }
-
-  get innerHTML() {
-    if (this._innerHTML) return this._innerHTML;
-    return this.childNodes.map((child) => child.outerHTML ?? child.textContent ?? '').join('');
-  }
-
-  set innerHTML(value) {
-    this._innerHTML = String(value);
-    this.childNodes = [];
-  }
-
-  appendChild(child) {
-    this._innerHTML = '';
-    return super.appendChild(child);
-  }
-
-  insertBefore(child, referenceNode) {
-    this._innerHTML = '';
-    return super.insertBefore(child, referenceNode);
-  }
-
-  removeChild(child) {
-    this._innerHTML = '';
-    return super.removeChild(child);
-  }
-
-  setAttribute(name, value) {
-    const stringValue = String(value);
-    this.attributes.set(name, stringValue);
-    if (name === 'class') {
-      this.className = stringValue;
-    } else if (name === 'id') {
-      this.id = stringValue;
-    } else if (name.startsWith('data-')) {
-      const key = name
-        .slice(5)
-        .split('-')
-        .map((part, index) => (index === 0 ? part : `${part[0]?.toUpperCase() ?? ''}${part.slice(1)}`))
-        .join('');
-      this.dataset[key] = stringValue;
-    }
-  }
-
-  getAttribute(name) {
-    return this.attributes.get(name) ?? null;
-  }
-
-  hasAttribute(name) {
-    return this.attributes.has(name);
-  }
-
-  removeAttribute(name) {
-    this.attributes.delete(name);
-    if (name === 'class') this.className = '';
-  }
-
-  matches(selector) {
-    return matchesSelector(this, selector);
-  }
-
-  querySelector(selector) {
-    return querySelectorAll(this, selector)[0] ?? null;
-  }
-
-  querySelectorAll(selector) {
-    return querySelectorAll(this, selector);
-  }
-
-  closest(selector) {
-    let current = this;
-    while (current instanceof MiniElement) {
-      if (current.matches(selector)) return current;
-      current = current.parentElement;
-    }
-    return null;
-  }
-
-  remove() {
-    if (this.parentNode) {
-      this.parentNode.removeChild(this);
-    }
-  }
-
-  getBoundingClientRect() {
-    return { width: 1, height: 1, top: 0, left: 0, right: 1, bottom: 1 };
-  }
-
-  focus() {
-    const doc = this.ownerDocument ?? globalThis.document;
-    if (doc) doc.activeElement = this;
-  }
-
-  get nextElementSibling() {
-    if (!this.parentNode) return null;
-    const siblings = this.parentNode.childNodes.filter((child) => child instanceof MiniElement);
-    const index = siblings.indexOf(this);
-    return index >= 0 ? siblings[index + 1] ?? null : null;
-  }
-
-  get isConnected() {
-    let current = this.parentNode;
-    while (current) {
-      if (current === globalThis.document?.body || current === globalThis.document?.documentElement) {
-        return true;
-      }
-      current = current.parentNode;
-    }
-    return false;
-  }
-
-  get outerHTML() {
-    return `<${this.tagName.toLowerCase()}>${this.innerHTML}</${this.tagName.toLowerCase()}>`;
-  }
-
-  get children() {
-    return this.childNodes.filter((child) => child instanceof MiniElement);
-  }
-
-  get offsetParent() {
-    return this.isConnected ? (this.parentElement ?? null) : null;
-  }
-}
-
-class MiniStorage {
-  constructor() {
-    this.values = new Map();
-  }
-
-  getItem(key) {
-    return this.values.has(key) ? this.values.get(key) : null;
-  }
-
-  setItem(key, value) {
-    this.values.set(key, String(value));
-  }
-
-  removeItem(key) {
-    this.values.delete(key);
-  }
-
-  clear() {
-    this.values.clear();
-  }
-}
-
-class MiniDocument extends EventTarget {
-  constructor() {
-    super();
-    this.documentElement = new MiniElement('html');
-    this.documentElement.clientHeight = 800;
-    this.documentElement.clientWidth = 1200;
-    this.body = new MiniElement('body');
-    this.documentElement.ownerDocument = this;
-    this.body.ownerDocument = this;
-    this.documentElement.appendChild(this.body);
-    this.activeElement = this.body;
-  }
-
-  createElement(tagName) {
-    const element = new MiniElement(tagName);
-    element.ownerDocument = this;
-    return element;
-  }
-
-  createTextNode(value) {
-    return new MiniText(value);
-  }
-
-  createDocumentFragment() {
-    return new MiniDocumentFragment();
-  }
-
-  getElementById(id) {
-    return querySelectorAll(this.documentElement, `#${id}`)[0] ?? null;
-  }
-
-  querySelector(selector) {
-    return this.documentElement.querySelector(selector);
-  }
-
-  querySelectorAll(selector) {
-    return this.documentElement.querySelectorAll(selector);
-  }
-}
-
-function splitSelectorList(selector) {
-  return String(selector)
-    .split(',')
-    .map((part) => part.trim())
-    .filter(Boolean);
-}
-
-function parseSimpleSelector(selector) {
-  const trimmed = selector.trim();
-  const result = {
-    tag: null,
-    id: null,
-    classes: [],
-    attributes: [],
-    notAttributes: [],
-  };
-  let remaining = trimmed;
-
-  const tagMatch = remaining.match(/^[a-zA-Z][a-zA-Z0-9-]*/);
-  if (tagMatch) {
-    result.tag = tagMatch[0].toUpperCase();
-    remaining = remaining.slice(tagMatch[0].length);
-  }
-
-  while (remaining.length > 0) {
-    if (remaining.startsWith('#')) {
-      const match = remaining.match(/^#([A-Za-z0-9_-]+)/);
-      if (!match) break;
-      result.id = match[1];
-      remaining = remaining.slice(match[0].length);
-      continue;
-    }
-
-    if (remaining.startsWith('.')) {
-      const match = remaining.match(/^\.([A-Za-z0-9_-]+)/);
-      if (!match) break;
-      result.classes.push(match[1]);
-      remaining = remaining.slice(match[0].length);
-      continue;
-    }
-
-    if (remaining.startsWith(':not(')) {
-      const match = remaining.match(/^:not\(\[([^\]=]+)(?:="([^"]*)")?\]\)/);
-      if (!match) break;
-      result.notAttributes.push({ name: match[1], value: match[2] ?? null });
-      remaining = remaining.slice(match[0].length);
-      continue;
-    }
-
-    if (remaining.startsWith('[')) {
-      const match = remaining.match(/^\[([^\]=]+)(?:="([^"]*)")?\]/);
-      if (!match) break;
-      result.attributes.push({ name: match[1], value: match[2] ?? null });
-      remaining = remaining.slice(match[0].length);
-      continue;
-    }
-
-    break;
-  }
-
-  return result;
-}
-
-function matchesSelector(element, selector) {
-  return splitSelectorList(selector).some((part) => {
-    const parsed = parseSimpleSelector(part);
-    if (parsed.tag && element.tagName !== parsed.tag) return false;
-    if (parsed.id && element.id !== parsed.id) return false;
-    if (parsed.classes.some((name) => !element.classList.contains(name))) return false;
-    if (parsed.attributes.some(({ name, value }) => {
-      if (!element.hasAttribute(name)) return true;
-      return value != null && element.getAttribute(name) !== value;
-    })) return false;
-    if (parsed.notAttributes.some(({ name, value }) => {
-      if (!element.hasAttribute(name)) return false;
-      return value == null ? true : element.getAttribute(name) === value;
-    })) return false;
-    return true;
-  });
-}
-
-function querySelectorAll(root, selector) {
-  const matches = [];
-
-  function visit(node) {
-    if (!(node instanceof MiniElement)) return;
-    if (node.matches(selector)) {
-      matches.push(node);
-    }
-    node.childNodes.forEach(visit);
-  }
-
-  if (root instanceof MiniElement) {
-    root.childNodes.forEach(visit);
-    return matches;
-  }
-
-  root.childNodes.forEach(visit);
-  return matches;
-}
-
-export function createBrowserEnvironment() {
-  const document = new MiniDocument();
-  const localStorage = new MiniStorage();
-  const window = {
-    document,
-    localStorage,
-    innerHeight: 800,
-    innerWidth: 1200,
-    addEventListener() {},
-    removeEventListener() {},
-    open() {},
-    location: {
-      origin: 'https://worldmonitor.test',
-      href: 'https://worldmonitor.test/',
-    },
-    navigator: {
-      clipboard: {
-        async writeText() {},
-      },
-    },
-    getComputedStyle() {
-      return {
-        display: '',
-        visibility: '',
-        gridTemplateColumns: 'none',
-        columnGap: '0',
-      };
-    },
-  };
-
-  return {
-    document,
-    localStorage,
-    window,
-    requestAnimationFrame(callback) {
-      if (typeof callback === 'function') callback(0);
-      return 1;
-    },
-    cancelAnimationFrame() {},
-    HTMLElement: MiniElement,
-    HTMLButtonElement: MiniElement,
-  };
-}
 
 function snapshotGlobal(name) {
   return {
