+++ tests/panel-mount-deferral.test.mts
@@ -2,157 +2,19 @@ import assert from 'node:assert/strict';
 import { readFile } from 'node:fs/promises';
 import { afterEach, describe, it } from 'node:test';
 
+import { createBrowserEnvironment } from './helpers/mini-dom.mts';
 import {
   countInteractiveControls,
   createDeferredPanelShell,
   getDeferredPanelShellFootprint,
   getInitialPanelMountBudget,
+  reconcileDeferredPanelShellColSpan,
   shouldDeferInitialPanelMount,
 } from '../src/app/panel-mount-deferral';
-class MiniClassList {
-  private values = new Set<string>();
-
-  add(...tokens: string[]): void {
-    for (const token of tokens) this.values.add(token);
-  }
-
-  remove(...tokens: string[]): void {
-    for (const token of tokens) this.values.delete(token);
-  }
-
-  contains(token: string): boolean {
-    return this.values.has(token);
-  }
-
-  setFromString(value: string): void {
-    this.values = new Set(String(value).split(/\s+/).filter(Boolean));
-  }
-
-  toString(): string {
-    return Array.from(this.values).join(' ');
-  }
-}
-
-class MiniText {
-  parentNode: MiniElement | null = null;
-  parentElement: MiniElement | null = null;
-
-  constructor(private value: string) {}
-
-  get textContent(): string {
-    return this.value;
-  }
-
-  set textContent(value: string | null) {
-    this.value = String(value ?? '');
-  }
-}
-
-class MiniElement {
-  readonly childNodes: Array<MiniElement | MiniText> = [];
-  readonly classList = new MiniClassList();
-  readonly attributes = new Map<string, string>();
-  readonly dataset: Record<string, string> = {};
-  parentNode: MiniElement | null = null;
-  parentElement: MiniElement | null = null;
-  id = '';
-
-  constructor(readonly tagName: string) {}
-
-  get className(): string {
-    return this.classList.toString();
-  }
-
-  set className(value: string) {
-    this.classList.setFromString(value);
-  }
-
-  get textContent(): string {
-    return this.childNodes.map((child) => child.textContent ?? '').join('');
-  }
-
-  set textContent(value: string | null) {
-    this.childNodes.length = 0;
-    this.appendChild(new MiniText(String(value ?? '')));
-  }
-
-  appendChild<T extends MiniElement | MiniText>(child: T): T {
-    if (child.parentNode) child.parentNode.removeChild(child);
-    child.parentNode = this;
-    child.parentElement = this;
-    this.childNodes.push(child);
-    return child;
-  }
-
-  removeChild<T extends MiniElement | MiniText>(child: T): T {
-    const index = this.childNodes.indexOf(child);
-    if (index >= 0) this.childNodes.splice(index, 1);
-    child.parentNode = null;
-    child.parentElement = null;
-    return child;
-  }
-
-  setAttribute(name: string, value: string): void {
-    const stringValue = String(value);
-    this.attributes.set(name, stringValue);
-    if (name === 'class') this.className = stringValue;
-    else if (name === 'id') this.id = stringValue;
-    else if (name.startsWith('data-')) {
-      const key = name.slice(5).replace(/-([a-z])/g, (_match, part: string) => part.toUpperCase());
-      this.dataset[key] = stringValue;
-    }
-  }
-
-  getAttribute(name: string): string | null {
-    return this.attributes.get(name) ?? null;
-  }
-
-  querySelector(selector: string): MiniElement | null {
-    return this.querySelectorAll(selector)[0] ?? null;
-  }
-
-  querySelectorAll(selector: string): MiniElement[] {
-    const selectors = selector.split(',').map((part) => part.trim()).filter(Boolean);
-    const matches: MiniElement[] = [];
-    const visit = (node: MiniElement) => {
-      for (const child of node.childNodes) {
-        if (!(child instanceof MiniElement)) continue;
-        if (selectors.some((part) => matchesSelector(child, part))) matches.push(child);
-        visit(child);
-      }
-    };
-    visit(this);
-    return matches;
-  }
-}
-
-class MiniDocument {
-  readonly body = new MiniElement('body');
-
-  createElement(tagName: string): MiniElement {
-    return new MiniElement(tagName.toLowerCase());
-  }
-}
-
-function matchesSelector(element: MiniElement, selector: string): boolean {
-  if (selector === '*') return true;
-  if (selector.startsWith('.')) return element.classList.contains(selector.slice(1));
-  if (selector === 'a[href]') return element.tagName === 'a' && element.attributes.has('href');
-  if (selector === '[tabindex]:not([tabindex="-1"])') return element.attributes.has('tabindex') && element.getAttribute('tabindex') !== '-1';
-  const role = selector.match(/^\[role="([^\"]+)"\]$/);
-  if (role) return element.getAttribute('role') === role[1];
-  return element.tagName === selector.toLowerCase();
-}
-
-function createBrowserEnvironment() {
-  return {
-    document: new MiniDocument(),
-    HTMLElement: MiniElement,
-  };
-}
 
 const originalDocument = Object.getOwnPropertyDescriptor(globalThis, 'document');
 const originalHTMLElement = Object.getOwnPropertyDescriptor(globalThis, 'HTMLElement');
+const originalWindow = Object.getOwnPropertyDescriptor(globalThis, 'window');
 
 function installDom() {
   const env = createBrowserEnvironment();
@@ -166,6 +28,11 @@ function installDom() {
     writable: true,
     value: env.HTMLElement,
   });
+  Object.defineProperty(globalThis, 'window', {
+    configurable: true,
+    writable: true,
+    value: env.window,
+  });
   return env.document;
 }
 
@@ -174,6 +41,8 @@ function restoreDom(): void {
   else delete (globalThis as { document?: unknown }).document;
   if (originalHTMLElement) Object.defineProperty(globalThis, 'HTMLElement', originalHTMLElement);
   else delete (globalThis as { HTMLElement?: unknown }).HTMLElement;
+  if (originalWindow) Object.defineProperty(globalThis, 'window', originalWindow);
+  else delete (globalThis as { window?: unknown }).window;
 }
 
 function createFullPanel(id: string): HTMLElement {
@@ -259,6 +128,31 @@ describe('panel mount deferral', () => {
     assert.equal(tallShell.classList.contains('span-2'), true);
   });
 
+  it('clamps saved deferred-shell column spans to the rendered grid width after insertion', () => {
+    const document = installDom();
+    const grid = document.createElement('div');
+    grid.className = 'panels-grid';
+    Object.defineProperty(grid, 'getBoundingClientRect', {
+      configurable: true,
+      value: () => ({ width: 560, height: 0, top: 0, left: 0, right: 560, bottom: 0, x: 0, y: 0, toJSON: () => ({}) }),
+    });
+    (globalThis.window as unknown as { getComputedStyle: () => { gridTemplateColumns: string; columnGap: string } }).getComputedStyle = () => ({
+      gridTemplateColumns: 'repeat(auto-fill, minmax(280px, 1fr))',
+      columnGap: '0',
+    });
+
+    const shell = createDeferredPanelShell(
+      'live-webcams',
+      'Live Webcams',
+      getDeferredPanelShellFootprint({ panelId: 'live-webcams', savedColSpans: { 'live-webcams': 3 } }),
+    );
+    grid.appendChild(shell);
+    reconcileDeferredPanelShellColSpan(shell);
+
+    assert.equal(shell.classList.contains('col-span-3'), false);
+    assert.equal(shell.classList.contains('col-span-2'), true);
+  });
+
   it('lets saved user spans override natural deferred-shell footprints', () => {
     const document = installDom();
     const footprint = getDeferredPanelShellFootprint({
@@ -320,7 +214,7 @@ describe('panel mount deferral', () => {
 
   it('signals queued panel work after replacing a deferred shell with the real panel', async () => {
     const source = await readFile(new URL('../src/app/panel-layout.ts', import.meta.url), 'utf8');
-    const mountPanelElement = source.match(/private\s+mountPanelElement[\s\S]*?\n  \}/);
+    const mountPanelElement = source.match(/private\s+mountPanelElement[\s\S]*?\n {2}\}/);
 
     assert.ok(mountPanelElement, 'mountPanelElement method not found');
     assert.match(
